Module Uspf_lwt

LWT layer of uSPF.

uSPF is a standalone library which requires some specialisations like the scheduler and the DNS stack used to get DNS record. This module specialises specifically the scheduler with LWT. It lets the user to choose the DNS implementation via a first-class module (see DNS).

NOTE: This layer is currently incomplete and it uses mostly use to get the DNS record and do some verifications on it - but it does not actually do (and can not) the SPF check (as the Unix layer, see Uspf_unix).

module type DNS = sig ... end

val get : domain:_ Domain_name.t -> 'dns -> (module DNS with type t = 'dns) -> (Uspf.Term.t, [> `Not_found | `Invalid_SPF_record ]) Stdlib.result Lwt.t